docs: renamed fake review to normal review
