+++ deno.json
@@ -6,7 +6,7 @@
     "@deco/codemod-toolkit": "jsr:@deco/codemod-toolkit@0.3.4",
     "@deco/durable": "jsr:@deco/durable@0.5.3",
     "@deco/warp": "jsr:@deco/warp@0.3.8",
-    "@eluxlab/library-components": "https://esm.sh/@eluxlab/library-components@0.3.5?external=react,react-dom,react/jsx-runtime&target=es2022",
+    "@eluxlab/library-components": "https://esm.sh/@eluxlab/library-components@0.3.6?external=react,react-dom,react/jsx-runtime&target=es2022",
     "@hono/hono": "jsr:@hono/hono@^4.5.4",
     "@std/assert": "jsr:@std/assert@^1.0.2",
     "@std/async": "jsr:@std/async@^0.224.1",
